+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Use of the deprecated method ReflectionProperty::setAccessible() with PHP > 8.0
+
 ## [1.2.0] - 2026-01-13
 
 ### Removed